Move unstable api into a separate mod
