+++ mp4parse_capi/tests/test_sample_table.rs
@@ -1,6 +1,6 @@
 extern crate mp4parse;
 extern crate mp4parse_capi;
-use mp4parse::Indice;
+use mp4parse::unstable::Indice;
 use mp4parse_capi::*;
 use std::io::Read;
